chore: update create project mutation
